pyproject.toml
[project]
name = "cli-stock-picker"
version = "1.0.0"
description = "Stock Intelligence Multi-Agent System using LangGraph"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=1.0.5",
    "langchain-core>=1.0.3",
    "langchain-openai>=1.1.0",
    "langgraph-checkpoint-sqlite>=2.0.2",
    "pydantic>=2.10.5",
    "yfinance>=0.2.50",
    "rich>=13.9.4",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
